+++ src/perpl/rot_2d_symm_fit.py
@@ -37,13 +37,14 @@ from tkinter.filedialog import askopenfilename
 import numpy as np
 import matplotlib.pyplot as plt
 
-from perpl.background_models import zero_to_constant_gradient
-import perpl.modelling_general as models
-from perpl.modelling_general import ModelWithFitSettings
-from perpl.modelling_general import stdev_of_model
-import perpl.modelstats as stats
+from perpl.io import plotting, reports
+from perpl.modelling.background_models import zero_to_constant_gradient
+import perpl.modelling.modelling_general as models
+from perpl.modelling.modelling_general import ModelWithFitSettings
+from perpl.modelling.modelling_general import stdev_of_model
+import perpl.statistics.modelstats as stats
 from perpl.relative_positions import getdistances
-from perpl import utils, plotting, reports
+from perpl.io import utils
 
 
 class Number:
